requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
executing==2.0.1
FLAML==2.1.1
fonttools==4.47.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipython==8.20.0
isodate==0.6.1
jedi==0.19.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.0.246
langsmith==0.0.82
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.8
numpy==1.26.3
openai==1.8.0
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
pure-eval==0.2.2
pyautogen==0.2.9
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
# pytz==2023.3.post1
# pywin32==306
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.17.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.4
azure-storage-blob==12.8.1
sentence-transformers==2.3.1
